+++ src/common/storage.js
