--- src/config.js
@@ -1,13 +1,0 @@
-seajs.config({
-	path:".",
-	alias:{
-		"app":"src/app",
-		"ajax": "src/mod/ajax",
-		"events": "src/mod/events",
-		"model": "src/mod/model",
-		"parser": "src/mod/parser",
-		"storage": "src/mod/storage",
-		"view": "src/mod/view",
-		"commonStorage": "src/mod/commonStorage"
-	}
-});
